Rename t_*.rb to test_*.rb
