+++ test/run-test.rb
@@ -11,7 +11,7 @@ $LOAD_PATH.unshift("#{base_path}/test")
 client_base_path = Pathname(__FILE__).dirname.parent.parent.expand_path
 $LOAD_PATH.unshift("#{client_base_path}/roma-ruby-client/lib")
 
-require 'roma-test-utils'
+require 'roma_test_utils'
 
 puts "* ROMA Version : #{Roma::Config::VERSION}"
 puts "* Ruby Client Version : #{Roma::Client::VERSION::STRING}"
@@ -21,7 +21,7 @@ Dir["#{base_path}/test/t_*.rb"].each do |test_file|
 end
 
 ### optional test
-require_relative './optional_test/t_routing_logic.rb'
+require_relative './optional_test/test_routing_logic.rb'
 #require_relative './optional_test/t_mkroute_rich.rb'
 #require_relative './optional_test/t_other_database.rb'
 #require_relative './optional_test/t_other_cpdb.rb'
