+++ test/test_command_definition.rb
