+++ test/test_cpdata.rb
