+++ test/test_cpdb.rb
