+++ test/test_eventmachine.rb
