+++ test/test_listplugin.rb
