+++ test/test_logshift.rb
