+++ test/test_mapcountplugin.rb
