+++ test/test_mapplugin.rb
