+++ test/test_mhash.rb
