+++ test/test_new_func.rb
