+++ test/test_rclient.rb
