+++ test/test_storage_error.rb
